+++ src/resources/event.rs
@@ -1,12 +1,11 @@
-use error::Error;
-use client::Client;
-use resources::{Charge, Invoice};
+use resources::{Charge, Invoice, Subscription};
 
 #[derive(Debug, PartialEq, Deserialize, Serialize)]
 pub enum EventType {
+    #[serde(rename = "charge.succeeded")] ChargeSucceeded,
+    #[serde(rename = "customer.subscription.created")] CustomerSubscriptionCreated,
     #[serde(rename = "invoice.created")] InvoiceCreated,
     #[serde(rename = "invoice.updated")] InvoiceUpdated,
-    #[serde(rename = "charge.succeeded")] ChargeSucceeded,
 }
 
 #[derive(Debug, Deserialize)]
@@ -27,4 +26,5 @@ pub struct EventData {
 pub enum EventObject {
     #[serde(rename = "charge")] Charge(Charge),
     #[serde(rename = "invoice")] Invoice(Invoice),
+    #[serde(rename = "subscription")] Subscription(Subscription),
 }