+++ src/resources/invoices.rs
@@ -20,6 +20,18 @@ pub struct InvoiceParams<'a> {
     #[serde(skip_serializing_if = "Option::is_none")] pub forgiven: Option<bool>,
 }
 
+#[derive(Default, Serialize)]
+pub struct InvoiceItemParams<'a> {
+    #[serde(skip_serializing_if = "Option::is_none")] pub amount: Option<i64>,
+    #[serde(skip_serializing_if = "Option::is_none")] pub currency: Option<Currency>,
+    #[serde(skip_serializing_if = "Option::is_none")] pub customer: Option<&'a str>,
+    #[serde(skip_serializing_if = "Option::is_none")] pub description: Option<&'a str>,
+    #[serde(skip_serializing_if = "Option::is_none")] pub discountable: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")] pub invoice: Option<&'a str>,
+    #[serde(skip_serializing_if = "Option::is_none")] pub metadata: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")] pub subscription: Option<bool>,
+}
+
 /*
 #[derive(Serialize)]
 pub struct InvoiceListLinesParams {
@@ -42,9 +54,9 @@ pub struct Period {
 ///
 /// For more details see https://stripe.com/docs/api#invoice_line_item_object.
 #[derive(Debug, Deserialize)]
-pub struct InvoiceLine {
+pub struct InvoiceItem {
     pub id: String,
-    pub amount: u64,
+    pub amount: i64,
     pub currency: Currency,
     pub description: Option<String>,
     pub discountable: bool,
@@ -56,7 +68,9 @@ pub struct InvoiceLine {
     pub quantity: Option<u64>,
     pub subscription: Option<String>,
     pub subscription_item: Option<String>,
-    #[serde(rename = "type")] pub item_type: String, // (invoiceitem, subscription)
+    #[serde(default)] // NOTE: Missing in response to InvoiceItem create
+    #[serde(rename = "type")]
+    pub item_type: String, // (invoiceitem, subscription)
 }
 
 /// The resource representing a Stripe invoice.
@@ -78,7 +92,7 @@ pub struct Invoice {
     pub discount: Option<Discount>,
     pub ending_balance: Option<i64>,
     pub forgiven: bool,
-    pub lines: List<InvoiceLine>,
+    pub lines: List<InvoiceItem>,
     pub livemode: bool,
     pub metadata: Metadata,
     pub next_payment_attempt: Option<Timestamp>,
@@ -119,7 +133,7 @@ impl Invoice {
     }
 
     // TODO: Implement InvoiceListLinesParams
-    // pub fn get_lines(client: &Client, invoice_id: &str, params: InvoiceListLinesParams) -> Result<List<InvoiceLine>, Error> {
+    // pub fn get_lines(client: &Client, invoice_id: &str, params: InvoiceListLinesParams) -> Result<List<InvoiceItem>, Error> {
     //     client.get(&format!("/invoices/{}/lines", invoice_id))
     // }
 
@@ -148,4 +162,14 @@ impl Invoice {
     pub fn list(client: &Client, params: InvoiceListParams) -> Result<List<Invoice>, Error> {
         client.get(&format!("/invoices?{}", query::to_string(&params)?))
     }
+}
+
+impl InvoiceItem {
+    /// Creates an invoice line item.
+    ///
+    /// For more details see https://stripe.com/docs/api/node#invoice_line_item_object
+    pub fn create(client: &Client, params: InvoiceItemParams) -> Result<InvoiceItem, Error> {
+        client.post(&format!("/invoiceitems"), &params)
+    }
+
 }